fix: added new relevance tokens. now the system allows null quotes json files. system tested until begining of third instruction. min_freq of profile domains hardcoded to 2 for testing reasons

+++ src/tsundoku/analysis/detect_anomalies.py
@@ -90,7 +90,7 @@ def main(experiment):
 
     user_data["feature.default_profile_image"] = user_data[
         "user.default_profile_image"
-    ].astype(np.int8)
+    ].astype(int)
 
     user_features = user_data.join(user_transformed_volume, on="user.id")
